+++ docs/guides/changelog.md
@@ -11,6 +11,7 @@ description: |-
 
 - Add `sdwan_system_flexible_port_speed_profile_parcel` resource and data source
 - Add `sdwan_system_security_profile_parcel` resource and data source
+- Add `sdwan_system_mrf_profile_parcel` resource and data source
 
 ## 0.3.7